Town Ruin Separation
- New settings class TownRuinSettings.
