+++ src/com/palmergames/bukkit/towny/war/siegewar/SiegeWarSettings.java
@@ -242,26 +242,6 @@ public class SiegeWarSettings {
 		return TownySettings.getInt(ConfigNodes.WAR_SIEGE_BANNER_CONTROL_SESSION_DURATION_MINUTES);
 	}
 
-	public static boolean getWarCommonTownRuinsEnabled() {
-		return TownySettings.getBoolean(ConfigNodes.WAR_COMMON_TOWN_RUINS_ENABLED);
-	}
-
-	public static int getWarCommonTownRuinsMaxDurationHours() {
-		return TownySettings.getInt(ConfigNodes.WAR_COMMON_TOWN_RUINS_MAX_DURATION_HOURS);
-	}
-
-	public static int getWarCommonTownRuinsMinDurationHours() {
-		return TownySettings.getInt(ConfigNodes.WAR_COMMON_TOWN_RUINS_MIN_DURATION_HOURS);
-	}
-
-	public static boolean getWarCommonTownRuinsReclaimEnabled() {
-		return TownySettings.getBoolean(ConfigNodes.WAR_COMMON_TOWN_RUINS_RECLAIM_ENABLED);
-	}
-
-	public static double getEcoPriceReclaimTown() {
-		return TownySettings.getDouble(ConfigNodes.ECO_PRICE_RECLAIM_RUINED_TOWN);
-	}
-
 	public static boolean getWarSiegePopulationBasedPointBoostsEnabled() {
 		return TownySettings.getBoolean(ConfigNodes.WAR_SIEGE_POPULATION_BASED_POINT_BOOSTS_ENABLED);
 	}
